requirements.txt
environs~=8.0.0
asyncpg~=0.27.0

aiohttp~=3.9.5
telebot~=0.0.5
mysql~=0.0.3
mysql-connector-python~=8.4.0
paho-mqtt~=2.1.0